+++ Beyond.Todo.Console/ITodoList.cs
@@ -5,7 +5,7 @@ public interface ITodoList
     void AddItem(string title, string description, string category);
     void UpdateItem(int id, string description);
     void RemoveItem(int id);
-    void RegisterProgression(int id, DateTime dateTime, decimal percent);
+    void RegisterProgression(int id, DateTime dateTime, int percent);
     void PrintItems();
     Task<List<string>> GetCategories();
 }
